move card generation to metadata
